Add parameter to keep elb rules
Does not purge elb rules. This is usefull if running the elb_application_lb
 role and there is the desire to keep existing rules.

+++ lib/ansible/modules/cloud/amazon/elb_application_lb.py
@@ -119,6 +119,10 @@ options:
     description:
       - The time in seconds to use in conjunction with I(wait).
     version_added: 2.6
+  keep_rules:
+    description:
+      - Keeps the existing load balancer rules in place. Will modify and add, but will not delete.
+    version_added: 2.7
 extends_documentation_fragment:
     - aws
     - ec2
@@ -444,10 +448,11 @@ def create_or_update_elb(elb_obj):
             rules_to_add, rules_to_modify, rules_to_delete = rules_obj.compare_rules()
 
             # Delete rules
-            for rule in rules_to_delete:
-                rule_obj = ELBListenerRule(elb_obj.connection, elb_obj.module, {'RuleArn': rule}, rules_obj.listener_arn)
-                rule_obj.delete()
-                elb_obj.changed = True
+            if not elb_obj.module.params['keep_rules']:
+                for rule in rules_to_delete:
+                    rule_obj = ELBListenerRule(elb_obj.connection, elb_obj.module, {'RuleArn': rule}, rules_obj.listener_arn)
+                    rule_obj.delete()
+                    elb_obj.changed = True
 
             # Add rules
             for rule in rules_to_add:
@@ -524,7 +529,8 @@ def main():
             state=dict(choices=['present', 'absent'], type='str'),
             tags=dict(type='dict'),
             wait_timeout=dict(type='int'),
-            wait=dict(default=False, type='bool')
+            wait=dict(default=False, type='bool'),
+            keep_rules=dict(default=False, type='bool')
         )
     )
